% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{attach_cluster_to_pixeldata}
\alias{attach_cluster_to_pixeldata}
\title{Attach clustering results back to Cardinal pixelData}
\usage{
attach_cluster_to_pixeldata(msi_obj, cluster_df)
}
\arguments{
\item{msi_obj}{A Cardinal MSI object.}

\item{cluster_df}{A data.frame containing at least:
\code{pixel_ID}, \code{UMAP1}, \code{UMAP2}, \code{kmeans_cluster}.}
}
\value{
The input MSI object with updated pixelData.
}
\description{
Attach clustering results back to Cardinal pixelData
}
